+++ tests/ContainerTest.php
@@ -9,17 +9,11 @@ use Innmind\DI\{
     Exception\ServiceNotFound,
     Exception\CircularDependency,
 };
-use Innmind\BlackBox\{
-    PHPUnit\BlackBox,
-    PHPUnit\Framework\TestCase,
-    Set,
-};
+use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 use Fixtures\Innmind\DI\Services;
 
 class ContainerTest extends TestCase
 {
-    use BlackBox;
-
     public function testInterface()
     {
         $this->assertInstanceOf(Container::class, Builder::new()->build());
@@ -27,78 +21,54 @@ class ContainerTest extends TestCase
 
     public function testConstructingTheDefinitionsIsImmutable()
     {
-        $this
-            ->forAll(Set\Unicode::strings())
-            ->then(function($name) {
-                $container = Builder::new();
-                $container2 = $container->add($name, static fn() => new \stdClass);
+        $container = Builder::new();
+        $container2 = $container->add(Services::name, static fn() => new \stdClass);
 
-                $this->assertInstanceOf(Builder::class, $container2);
-                $this->assertNotSame($container2, $container);
-                $this->assertInstanceOf(\stdClass::class, $container2->build()($name));
+        $this->assertInstanceOf(Builder::class, $container2);
+        $this->assertNotSame($container2, $container);
+        $this->assertInstanceOf(\stdClass::class, $container2->build()(Services::name));
 
-                try {
-                    $container->build()($name);
-                    $this->fail('it should throw');
-                } catch (\Exception $e) {
-                    $this->assertInstanceOf(ServiceNotFound::class, $e);
-                    $this->assertSame($name, $e->getMessage());
-                }
-            });
+        try {
+            $container->build()(Services::name);
+            $this->fail('it should throw');
+        } catch (\Exception $e) {
+            $this->assertInstanceOf(ServiceNotFound::class, $e);
+            $this->assertSame('Fixtures\Innmind\DI\Services::name', $e->getMessage());
+        }
     }
 
     public function testServiceIsOnlyBuiltOnce()
     {
-        $this
-            ->forAll(Set\Unicode::strings())
-            ->then(function($name) {
-                $container = Builder::new()
-                    ->add($name, static fn() => new \stdClass)
-                    ->build();
+        $container = Builder::new()
+            ->add(Services::name, static fn() => new \stdClass)
+            ->build();
 
-                $this->assertSame($container($name), $container($name));
-            });
+        $this->assertSame($container(Services::name), $container(Services::name));
     }
 
     public function testDependenciesCanBeAccesedWhenBuildingService()
     {
-        $this
-            ->forAll(
-                Set\Unicode::strings(),
-                Set\Unicode::strings(),
-            )
-            ->filter(static fn($a, $b) => $a !== $b)
-            ->then(function($name, $dependency) {
-                $container = Builder::new()
-                    ->add($name, static fn($get) => $get($dependency))
-                    ->add($dependency, static fn() => new \stdClass)
-                    ->build();
+        $container = Builder::new()
+            ->add(Services::name, static fn($get) => $get(Services::dependency))
+            ->add(Services::dependency, static fn() => new \stdClass)
+            ->build();
 
-                $this->assertSame($container($dependency), $container($name));
-            });
+        $this->assertSame($container(Services::dependency), $container(Services::name));
     }
 
     public function testCircularDependenciesAreIntercepted()
     {
-        $this
-            ->forAll(
-                Set\Unicode::strings(),
-                Set\Unicode::strings(),
-            )
-            ->filter(static fn($a, $b) => $a !== $b)
-            ->then(function($name, $dependency) {
-                $container = Builder::new()
-                    ->add($name, static fn($get) => $get($dependency))
-                    ->add($dependency, static fn($get) => $get($name))
-                    ->build();
+        $container = Builder::new()
+            ->add(Services::name, static fn($get) => $get(Services::dependency))
+            ->add(Services::dependency, static fn($get) => $get(Services::name))
+            ->build();
 
-                try {
-                    $container($name);
-                    $this->fail('it should throw');
-                } catch (CircularDependency $e) {
-                    $this->assertSame("$name > $dependency > $name", $e->getMessage());
-                }
-            });
+        try {
+            $container(Services::name);
+            $this->fail('it should throw');
+        } catch (CircularDependency $e) {
+            $this->assertSame("Fixtures\Innmind\DI\Services::name > Fixtures\Innmind\DI\Services::dependency > Fixtures\Innmind\DI\Services::name", $e->getMessage());
+        }
     }
 
     public function testEnumCaseCanBeUsedToReferenceAService()
@@ -109,5 +79,21 @@ class ContainerTest extends TestCase
             ->build();
 
         $this->assertSame($expected, $container(Services::a));
+    }
+
+    public function testServicesAreFreedFromMemoryWhenUnused()
+    {
+        $called = 0;
+        $container = Builder::new()
+            ->add(Services::name, static function() use (&$called) {
+                ++$called;
+
+                return new \stdClass;
+            })
+            ->build();
+
+        $container(Services::name);
+        $this->assertInstanceOf(\stdClass::class, $container(Services::name));
+        $this->assertSame(2, $called);
     }
 }