+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 3.0.0 - 2025-08-23
+
+### Changed
+
+- Unused services are freed from memory
+
+### Removed
+
+- Using a `string` as a service name
+
+### Fixed
+
+- Service names in exception messages that used object hashes
+
 ## 2.1.0 - 2024-03-24
 
 ### Added